requirements.txt
langgraph
openai
python-dotenv
tiktoken
pandas
scipy
matplotlib
langchain_openai